completed project

+++ graphql/resolvers/bookings.js
@@ -8,7 +8,7 @@ module.exports = {
       throw new Error("Unauthenticated!");
     }
     try {
-      const bookings = await Booking.find();
+      const bookings = await Booking.find({ user: req.userId });
       return bookings.map((booking) => {
         return transformBooking(booking);
       });
@@ -36,6 +36,9 @@ module.exports = {
       const booking = await Booking.findById({ _id: args.bookingId }).populate(
         "event"
       );
+      if (req.userId != booking.user._id) {
+        throw new Error("You haven't booked this event");
+      }
       const event = transformEvent(booking.event);
       await Booking.deleteOne({ _id: args.bookingId });
       return event;
